backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
--extra-index-url https://aiinfra.pkgs.visualstudio.com/PublicPackages/_packaging/onnxruntime-cuda-12/pypi/simple/

# Core dependencies
fastapi==0.115.6
uvicorn==0.30.0
pydantic==2.7.0
python-dotenv>=1.0.0
websockets>=12.0
requests>=2.31.0
python-multipart>=0.0.6
aiohttp>=3.0.0

# AI/ML dependencies
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1  # Compatible with torch 2.3.1
numpy<2.0,>=1.21.0
openai-whisper==20231117
faster-whisper>=1.0.0
transformers==4.51.3
diffusers==0.29.0
modelscope==1.20.0

# Audio processing
librosa==0.10.2
soundfile==0.12.1
pydub>=0.25.1
pyworld==0.3.4

# CosyVoice dependencies
gradio==5.4.0
omegaconf==2.3.0
hydra-core==1.3.2
HyperPyYAML==1.2.2
inflect==7.3.1
lightning==2.2.4
networkx==3.1
protobuf==4.25
pyarrow==18.1.0
rich==13.7.1
wetext==0.0.4

# ONNX Runtime (platform-specific)
# Use CPU version for Windows/macOS, GPU for Linux with CUDA 12
onnx==1.16.0
onnxruntime==1.18.0; sys_platform == 'darwin' or sys_platform == 'win32'
onnxruntime-gpu==1.18.0; sys_platform == 'linux'

# Linux-only dependencies
deepspeed==0.15.1; sys_platform == 'linux'
tensorrt-cu12==10.13.3.9; sys_platform == 'linux'
tensorrt-cu12-bindings==10.13.3.9; sys_platform == 'linux'
tensorrt-cu12-libs==10.13.3.9; sys_platform == 'linux'

# Optional/Utilities
gdown==5.1.0
wget==3.2
matplotlib==3.7.5
tensorboard==2.14.0
grpcio==1.57.0
grpcio-tools==1.57.0
ruamel.yaml<0.18.0,>=0.17.0
openai>=1.0.0
fastapi-cli==0.0.4
x-transformers==2.11.24